requirements.txt
streamlit>=1.33,<2
google-generativeai>=0.7.2,<1
